Update example to use viewportChangeOptions
Nothing big, but as a beginner, took a bit of time for me to look for how to actually change the flyTo options. It might be helpful to include in the documentation an example of how to use viewportChangeOptions as a prop, especially for more MapBox beginners.

Let me know if this makes sense! Thanks so much, am a big fan of your work and a constant user of this repo :D